Look for breaks in sequences of time blocks

Developed to help work with fishing regulation data. consolidate_time_blocks takes any number of rows, with a column of start and end dates, and consolidates all contiguous periods of dates into a single one (and so each row of the output represents a period with one or more days break with the last one). Moreover, this function uses the clipr package to use the system clipboard. So you can copy a block of dates from excel (like in the screenshot) and then run the function with no arguments to have it pull in the copied cells and run the calculations.

Usage

consolidate_time_blocks(.data = NULL, start_col = "start", end_col = "end")

Arguments

.data

Dataframe. Optional; if not provided, instead use system clipboard (assumes it contains a copy of a 2-column excel block)

start_col

Name of column containing start dates. Defaults to "start". Unused if .data is not provided.

end_col

Name of column containing end dates. Defaults to "start". Unused if .data is not provided.

Value

Tibble with start and end date of each non-contiguous period, and length in days

Examples

library(tidyr)
#> Error in library(tidyr): there is no package called ‘tidyr’
df = tribble(
~ start, ~end_date,
"2024-8-1", "2024-8-5",
"2024-8-6", "2024-8-10",
"2024-8-16", "2024-8-30"
)
#> Error in tribble(~start, ~end_date, "2024-8-1", "2024-8-5", "2024-8-6",     "2024-8-10", "2024-8-16", "2024-8-30"): could not find function "tribble"
df |> consolidate_time_blocks(end_col = "end_date") #column name for start column is the default, "start"
#> ! If dates are in ambiguous format, making best guess. Check years of output!
#> Error in .data[[start_col]]: object of type 'closure' is not subsettable